+++ packages/component-library/src/components/Card/Card.tsx
@@ -15,6 +15,7 @@ import { MediaProps } from '../Media/Media.types';
 import Link, { LinkProps } from '../Link';
 import Text, { RichText } from '../Text';
 import sidekick from '../../utils/sidekick';
+import ConditionalWrapper from '../ConditionalWrapper';
 
 export interface CardProps extends MuiCardProps {
   __typename: string;
@@ -24,66 +25,75 @@ export interface CardProps extends MuiCardProps {
   subtitle?: string;
   media?: MediaProps | MediaProps[];
   body?: RichText;
+  link?: LinkProps;
   actions?: LinkProps[];
   sidekickLookup: any;
 }
 
 export interface CardOverrides {}
 
-export const Card = ({ media, title, subtitle, body, actions, variant, loading, sidekickLookup }: CardProps) => {
+export const Card = ({ media, title, subtitle, body, link, actions, variant, loading, sidekickLookup }: CardProps) => {
   return (
     <ErrorBoundary>
-      <Root variant={variant} data-testid="Card" {...sidekick(sidekickLookup)}>
-        {media || loading ? (
-          <Box display="flex" justifyContent="center">
-            {loading ? (
-              <Media {...sidekick(sidekickLookup?.media)} {...(Array.isArray(media) ? media[0] : media)} />
-            ) : (
-              <Skeleton>
+      <ConditionalWrapper
+        condition={!!link}
+        wrapper={(children: any) => (
+          <Link noLinkStyle {...link}>
+            {children}
+          </Link>
+        )}>
+        <Root variant={variant} data-testid="Card" {...sidekick(sidekickLookup)}>
+          {media || loading ? (
+            <Box display="flex" justifyContent="center">
+              {!loading ? (
                 <Media {...sidekick(sidekickLookup?.media)} {...(Array.isArray(media) ? media[0] : media)} />
-              </Skeleton>
-            )}
-          </Box>
-        ) : null}
-        {!loading && (title || subtitle || body || actions) ? (
-          <CardContent>
-            {title ? (
-              <Typography {...sidekick(sidekickLookup?.title)} variant="h3" component="h3" data-testid="Card-title">
-                {title}
+              ) : (
+                <Skeleton>
+                  <Media {...sidekick(sidekickLookup?.media)} {...(Array.isArray(media) ? media[0] : media)} />
+                </Skeleton>
+              )}
+            </Box>
+          ) : null}
+          {!loading && (title || subtitle || body || actions) ? (
+            <CardContent>
+              {title ? (
+                <Typography {...sidekick(sidekickLookup?.title)} variant="h3" component="h3" data-testid="Card-title">
+                  {title}
+                </Typography>
+              ) : null}
+              {subtitle ? (
+                <Typography {...sidekick(sidekickLookup?.subtitle)} variant="h4" component="h4" data-testid="Card-subtitle">
+                  {subtitle}
+                </Typography>
+              ) : null}
+              {body ? <Text sidekickLookup={sidekickLookup?.body} body={body} data-testid="Card-body" /> : null}
+              {actions?.length ? (
+                <CardActions {...sidekick(sidekickLookup?.actions)} data-testid="Card-actions">
+                  {actions?.map((link) => (
+                    <Link key={link.id} {...link} />
+                  ))}
+                </CardActions>
+              ) : null}
+            </CardContent>
+          ) : null}
+          {loading ? (
+            <CardContent>
+              <Typography variant="h3" component="h3">
+                <Skeleton width="100%" />
               </Typography>
-            ) : null}
-            {subtitle ? (
-              <Typography {...sidekick(sidekickLookup?.subtitle)} variant="h4" component="h4" data-testid="Card-subtitle">
-                {subtitle}
+              <Typography variant="h4" component="h4">
+                <Skeleton width="100%" />
+                <br />
+                <Skeleton width="100%" />
               </Typography>
-            ) : null}
-            {/* {body ? <Text sidekickLookup={sidekickLookup?.body} body={body} /> : null} */}
-            TEXT
-            {actions?.length ? (
-              <CardActions {...sidekick(sidekickLookup?.actions)}>
-                {actions?.map((link) => (
-                  <Link key={link.id} {...link} />
-                ))}
+              <Skeleton>{body ? <Text sidekickLookup={sidekickLookup?.body} body={body} data-testid="Card-body" /> : null}</Skeleton>
+              <CardActions>
+                <Skeleton width={50} />
               </CardActions>
-            ) : null}
-          </CardContent>
-        ) : (
-          <CardContent>
-            <Typography variant="h3" component="h3">
-              <Skeleton width="100%" />
-            </Typography>
-            <Typography variant="h4" component="h4">
-              <Skeleton width="100%" />
-              <br />
-              <Skeleton width="100%" />
-            </Typography>
-            <Skeleton>{body ? <Text sidekickLookup={sidekickLookup?.body} body={body} /> : null}</Skeleton>
-            <CardActions>
-              <Skeleton width={50} />
-            </CardActions>
-          </CardContent>
-        )}
-      </Root>
+            </CardContent>
+          ) : null}
+        </Root>
+      </ConditionalWrapper>
     </ErrorBoundary>
   );
 };