+++ packages/component-library/src/components/Collection/Collection.mock.ts
@@ -2,6 +2,7 @@ import mockCard from '../Card/Card.mock';
 import mockTheme from '../../theme/mock.theme';
 
 export default {
+  itemsSpacing: 2,
   variant: 'collection-three-per-row',
   items: [{ ...mockCard }, { ...mockCard }, { ...mockCard }, { ...mockCard }, { ...mockCard }],
   itemsVariant: 'standard-round',